feat(app): add car page
move the list of available cars to sell to the car page
